+++ src/traits/delete/delete_interface.rs
@@ -3,7 +3,7 @@ use anyhow::Result;
 
 pub trait DeleteInterface<T: Table>: Send + Sync
 {
-    fn delete_by_id(&self, id: u32) -> Result<()>;
+    fn delete_by_id(&self, id: u64) -> Result<()>;
 
     fn delete_by_expression(&self, expression: &str) -> Result<u64>;
 }
